requirements.txt
Click==7.0
Flask==1.0.2
Jinja2==2.10
MarkupSafe==1.1.0
Werkzeug==0.14.1
asn1crypto==0.24.0
cffi==1.11.5
chardet==3.0.4
cryptography==2.5
idna==2.8
itsdangerous==1.1.0
pip==10.0.1
pycparser==2.19
requests==2.21
setuptools==39.1.0
six==1.12.0
urllib3==1.24.1
configparser==3.5.1
aiml==0.9.2
PyJWT==1.7.1
pyOpenSSL==19.0